+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # changelog
 
 * add `SliceIntersect` function in `basicnew` package
+* fix unexpected result for `SimilarSlice` in `basiccheck` when there are elements of arguments present multiple times, but not in the same way in each argument
 
 ## v0.12.0